Do not generate ref before elem_ptr and field_ptr
Previously AstGen would emit a Zir `ref` instruction before any `elem_ptr`
and `field_ptr` instruction. This instruction is now omitted, and the
relevant instructions in Sema are updated to account for this change.

analyzeRef is modified such that it now stores intermediates to a mutable
pointer, after which the pointer is casted to a constant version.