Fix order dependence of LibExeObjStep.install()
Instead of configuring the paths of a ArtifactInstallStep when it is
created, defer this until the make() call of the associated
LibExeObjStep. This means that any steps that depend on the paths of the
installed files (even if they do not need the files to be installed)
should depend on the LibExeObjStep and do any processing with the
installation paths after the LibExeObjStep has been made.