Builder: add build_args and addBuild
This change adds the "build_args" field to the zig Builder. This provide access to the command-line arguments that were passed to the current "zig build" instance.  These arguments are forwarded to the build_runner after a fixed set of arguments passed from zig itelf.  Builds no longer need to know this internal detail to access these arguments.

Along with this, I've also added a convenient "addBuild" function to Builder that will setup a RunStep to execute "zig build".