Handle EINVAL in std.os.readlink
Both readlink and readlinkat return EINVAL if bufsiz is not positive, which is in fact impossible (which I believe is why `unreachable` was originally put) however it also returns EINVAL in the case that the named file is not a symbolic link. This is an error that should be handled like file not found or not dir errors are.

So, this commit returns an `error.NotSymLink` when EINVAL is returned (and since bufsiz < 0 is in fact not a possible state, it's the only case where EINVAL is returned).

Handling this error correctly also fixes a bug in `std.zig.system.NativeTargetInfo.glibcVerFromSO` which caused the Zig compiler, when built from source on some Linux environments, to crash with a `reached unreachable code` error.