C backend: Handle NaN constants
Previously, they emited `nan` as constant, which wasn't valid as `math.h` wasn't included (and even if it was, the correct macro name is `NAN`).

I tried to find the simplest way of getting NaN in a way that doesn't require using math.h and that will always work as a constant in ISO C.
First, using 0.0f / 0.0f looks like a solution but it generates a trap which is not intended behaviour when accessing a constant, if we actually put it in a constant variable, then now it breaks depending on the compiler.
Second solution was to take `std.math.nan_u32` and bit-cast it to a float. However bitcasting it would have required `memcpy` which also makes it not a constant.
Overall, it is just better to stick to `NAN`which was standardized in <math.h> in ISO C99 with a fallback to `0.0f / 0.0f`.